+++ test/unit/test_contracts_graph_parsed.py
@@ -76,6 +76,7 @@ def populated_node_config_dict():
         'tags': [],
         'vars': {},
         'extra': 'even more',
+        'on_schema_change': 'ignore',
     }
 
 
@@ -154,6 +155,7 @@ def base_parsed_model_dict():
             'quoting': {},
             'tags': [],
             'vars': {},
+            'on_schema_change': 'ignore'
         },
         'deferred': False,
         'docs': {'show': True},
@@ -244,6 +246,7 @@ def complex_parsed_model_dict():
             'quoting': {},
             'tags': [],
             'vars': {'foo': 100},
+            'on_schema_change': 'ignore'
         },
         'docs': {'show': True},
         'columns': {
@@ -439,6 +442,7 @@ def basic_parsed_seed_dict():
             'quoting': {},
             'tags': [],
             'vars': {},
+            'on_schema_change': 'ignore'
         },
         'deferred': False,
         'docs': {'show': True},
@@ -532,6 +536,7 @@ def complex_parsed_seed_dict():
             'tags': [],
             'vars': {},
             'quote_columns': True,
+            'on_schema_change': 'ignore'
         },
         'deferred': False,
         'docs': {'show': True},
@@ -784,6 +789,7 @@ def base_parsed_hook_dict():
             'quoting': {},
             'tags': [],
             'vars': {},
+            'on_schema_change': 'ignore'
         },
         'docs': {'show': True},
         'columns': {},
@@ -854,6 +860,7 @@ def complex_parsed_hook_dict():
             'quoting': {},
             'tags': [],
             'vars': {},
+            'on_schema_change': 'ignore'
         },
         'docs': {'show': True},
         'columns': {
@@ -1000,7 +1007,8 @@ def basic_parsed_schema_test_dict():
             'schema': 'dbt_test__audit',
             'warn_if': '!= 0',
             'error_if': '!= 0',
-            'fail_calc': 'count(*)'
+            'fail_calc': 'count(*)',
+            'on_schema_change': 'ignore',
         },
         'docs': {'show': True},
         'columns': {},
@@ -1078,7 +1086,8 @@ def complex_parsed_schema_test_dict():
             'warn_if': '!= 0',
             'error_if': '!= 0',
             'fail_calc': 'count(*)',
-            'extra_key': 'extra value'
+            'extra_key': 'extra value',
+            'on_schema_change': 'ignore',
         },
         'docs': {'show': False},
         'columns': {
@@ -1198,6 +1207,7 @@ def basic_timestamp_snapshot_config_dict():
         'updated_at': 'last_update',
         'target_database': 'some_snapshot_db',
         'target_schema': 'some_snapshot_schema',
+        'on_schema_change': 'ignore',
     }
 
 
@@ -1230,6 +1240,7 @@ def complex_timestamp_snapshot_config_dict():
         'extra': 'even more',
         'strategy': 'timestamp',
         'updated_at': 'last_update',
+        'on_schema_change': 'ignore',
     }
 
 
@@ -1286,6 +1297,7 @@ def basic_check_snapshot_config_dict():
         'unique_key': 'id',
         'strategy': 'check',
         'check_cols': 'all',
+        'on_schema_change': 'ignore',
     }
 
 
@@ -1318,6 +1330,7 @@ def complex_set_snapshot_config_dict():
         'extra': 'even more',
         'strategy': 'check',
         'check_cols': ['a', 'b'],
+        'on_schema_change': 'ignore',
     }
 
 
@@ -1423,6 +1436,7 @@ def basic_timestamp_snapshot_dict():
             'unique_key': 'id',
             'strategy': 'timestamp',
             'updated_at': 'last_update',
+            'on_schema_change': 'ignore'
         },
         'docs': {'show': True},
         'columns': {},
@@ -1554,6 +1568,7 @@ def basic_check_snapshot_dict():
             'unique_key': 'id',
             'strategy': 'check',
             'check_cols': 'all',
+            'on_schema_change': 'ignore'
         },
         'docs': {'show': True},
         'columns': {},